pyproject.toml
[project]
name = "prosper-challenge"
version = "0.1.0"
description = "Template repository for clinic appointment scheduling voice agent challenge"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[webrtc,daily,silero,elevenlabs,openai,local-smart-turn-v3,runner]",
    "pipecat-ai-cli",
    "playwright",
    "python-dateutil>=2.9",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "pytest>=8",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]